# clients/shared/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Fatoura API",
    "description": "UAE VAT sales, purchases and invoicing.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/invoices": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "createdById",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/InvoiceStatus"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "unpaidOnly",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultOfInvoiceSummaryDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Invoices"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IssueInvoiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/invoices/{id}": {
      "get": {
        "tags": [
          "Invoices"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Invoices"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvoiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/invoices/{id}/void": {
      "post": {
        "tags": [
          "Invoices"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VoidInvoiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/invoices/{id}/payments": {
      "post": {
        "tags": [
          "Invoices"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/invoices/{id}/payments/{paymentId}": {
      "delete": {
        "tags": [
          "Invoices"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "paymentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvoiceDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyDto": {
        "required": [
          "name",
          "address",
          "phone",
          "email",
          "website",
          "trn"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "trn": {
            "type": "string"
          }
        }
      },
...
      "DiscountKind": {
        "enum": [
          "None",
          "Amount",
          "Percent"
        ]
      },
      "DocumentCurrencyDto": {
        "required": [
          "code",
          "exchangeRate",
          "subTotalAed",
          "vatTotalAed",
          "totalAed"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "exchangeRate": {
            "type": "number",
            "format": "double"
          },
          "subTotalAed": {
            "type": "number",
            "format": "double"
          },
          "vatTotalAed": {
            "type": "number",
            "format": "double"
          },
          "totalAed": {
            "type": "number",
            "format": "double"
          }
        },
        "description": "A document's currency, its rate (AED per unit) and its totals in AED, which reports and the VAT return use."
      },
      "DocumentDiscountDto": {
        "required": [
          "kind",
          "value",
          "amount"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/DiscountKind"
          },
          "value": {
            "type": "number",
            "format": "double"
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        },
        "description": "The discount as entered, and `Amount`, its value in money. The sub total before discount is `SubTotal + Amount`."
      },
      "DocumentDiscountRequest": {
        "required": [
          "kind",
          "value"
        ],
        "type": "object",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/DiscountKind"
          },
          "value": {
            "type": "number",
            "format": "double"
          }
        },
        "description": "A document-level discount: `Kind` Amount (AED) or Percent (of the sub total); omit or use None for no discount."
      },
...
      "DocumentLineRequest": {
        "required": [
          "itemId",
          "description",
          "quantity",
          "unitPrice",
          "taxCategory"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "description": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "unitPrice": {
            "type": "number",
            "format": "double"
          },
          "taxCategory": {
            "$ref": "#/components/schemas/TaxCategory"
          }
        },
        "description": "Validated by Task DocumentLines.ValidateAsync(IReadOnlyList&lt;DocumentLineRequest&gt;? lines, FatouraDbContext db, FieldValidator v, CancellationToken ct, bool allowInactiveItems = false) so all line errors are reported together."
      },
...
      "InvoiceDto": {
        "required": [
          "id",
          "number",
          "date",
          "clientId",
          "client",
          "company",
          "quotationId",
          "quotationNumber",
          "status",
          "voidReason",
          "voidedAt",
          "terms",
          "discount",
          "currency",
          "subTotal",
          "vatTotal",
          "total",
          "paidTotal",
          "creditedTotal",
          "balance",
          "createdById",
          "createdByName",
          "createdAt",
          "lines",
          "payments",
          "creditNotes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "client": {
            "$ref": "#/components/schemas/PartyDto"
          },
          "company": {
            "$ref": "#/components/schemas/CompanyDto"
          },
          "quotationId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "quotationNumber": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "voidReason": {
            "type": "string"
          },
          "voidedAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "terms": {
            "$ref": "#/components/schemas/TermsDto"
          },
          "discount": {
            "$ref": "#/components/schemas/DocumentDiscountDto"
          },
          "currency": {
            "$ref": "#/components/schemas/DocumentCurrencyDto"
          },
          "subTotal": {
            "type": "number",
            "format": "double"
          },
          "vatTotal": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "paidTotal": {
            "type": "number",
            "format": "double"
          },
          "creditedTotal": {
            "type": "number",
            "format": "double"
          },
          "balance": {
            "type": "number",
            "format": "double"
          },
          "createdById": {
            "type": "string",
            "format": "uuid"
          },
          "createdByName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceLineDto"
            }
          },
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentDto"
            }
          },
          "creditNotes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreditNoteSummaryDto"
            }
          }
        }
      },
...
      "InvoiceResult": {
        "required": [
          "invoice",
          "warnings"
        ],
        "type": "object",
        "properties": {
          "invoice": {
            "$ref": "#/components/schemas/InvoiceDto"
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockWarning"
            }
          }
        }
      },
      "InvoiceStatus": {
        "enum": [
          "Issued",
          "Void"
        ]
      },
      "InvoiceSummaryDto": {
        "required": [
          "id",
          "number",
          "date",
          "clientId",
          "clientName",
          "status",
          "currency",
          "total",
          "paidTotal",
          "creditedTotal",
          "balance",
          "createdByName"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "number": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "clientName": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "currency": {
            "type": "string"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "paidTotal": {
            "type": "number",
            "format": "double"
          },
          "creditedTotal": {
            "type": "number",
            "format": "double"
          },
          "balance": {
            "type": "number",
            "format": "double"
          },
          "createdByName": {
            "type": "string"
          }
        }
      },
      "IssueInvoiceRequest": {
        "required": [
          "clientId",
          "date",
          "lines",
          "terms",
          "payment"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentLineRequest"
            }
          },
          "terms": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TermsRequest"
              }
            ]
          },
          "payment": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/PaymentRequest"
              }
            ]
          },
          "discount": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentDiscountRequest"
              }
            ]
          },
          "currency": {
            "maxLength": 3,
            "type": [
              "null",
              "string"
            ]
          },
          "exchangeRate": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          }
        }
      },
...
      "PagedResultOfInvoiceSummaryDto": {
        "required": [
          "items",
          "total",
          "page",
          "pageSize"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceSummaryDto"
            }
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PartyDto": {
        "required": [
          "name",
          "address",
          "phone",
          "email",
          "trn"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "trn": {
            "type": "string"
          }
        }
      },
...
      "PaymentMethod": {
        "enum": [
          "Cash",
          "Card",
          "BankTransfer",
          "Cheque"
        ]
      },
      "PaymentRequest": {
        "required": [
          "amount",
          "method",
          "date",
          "reference"
        ],
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "double"
          },
          "method": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "date": {
            "type": [
              "null",
              "string"
            ],
            "format": "date"
          },
          "reference": {
            "maxLength": 100,
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "StockWarning": {
        "required": [
          "itemId",
          "itemName",
          "stockAfter"
        ],
        "type": "object",
        "properties": {
          "itemId": {
            "type": "integer",
            "format": "int32"
          },
          "itemName": {
            "type": "string"
          },
          "stockAfter": {
            "type": "number",
            "format": "double"
          }
        }
      },
...
      "TaxCategory": {
        "enum": [
          "Standard",
          "ZeroRated",
          "Exempt"
        ]
      },
      "TermsDto": {
        "required": [
          "paymentTerms",
          "completionOfWork",
          "notes",
          "closingText"
        ],
        "type": "object",
        "properties": {
          "paymentTerms": {
            "type": "string"
          },
          "completionOfWork": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "closingText": {
            "type": "string"
          }
        }
      },
      "TermsRequest": {
        "required": [
          "paymentTerms",
          "completionOfWork",
          "notes",
          "closingText"
        ],
        "type": "object",
        "properties": {
          "paymentTerms": {
            "maxLength": 1000,
            "type": [
              "null",
              "string"
            ]
          },
          "completionOfWork": {
            "maxLength": 1000,
            "type": [
              "null",
              "string"
            ]
          },
          "notes": {
            "maxLength": 4000,
            "type": [
              "null",
              "string"
            ]
          },
          "closingText": {
            "maxLength": 1000,
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Optional per-document terms; omitted values fall back to the defaults in Settings."
      },
      "UpdateInvoiceRequest": {
        "required": [
          "clientId",
          "date",
          "lines",
          "terms"
        ],
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentLineRequest"
            }
          },
          "terms": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TermsRequest"
              }
            ]
          },
          "discount": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentDiscountRequest"
              }
            ]
          },
          "currency": {
            "maxLength": 3,
            "type": [
              "null",
              "string"
            ]
          },
          "exchangeRate": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          }
        },
        "description": "Omitting the currency or rate keeps the invoice's own."
      },
...
      "VoidInvoiceRequest": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "maxLength": 500,
            "type": "string"
          }
        }
      }
    }
  },
...
}